Added scale change. Added color sketches.
Images are represented as an image pyramid of multiple resolutions (do not
make the scale change too much). And in the E step, the template is allowed to perturb its scale over the defined resolutions.

For illustration, added one more sketched image, where different templates
(clusters) are rendered with different colors).